requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
pyjwt==2.8.0
pillow==10.0.0
pydicom==2.4.2
numpy==1.26.0
matplotlib==3.8.1
# ML/deep learning - optional here; for training put them in model/requirements.txt
torch>=1.13.1
torchvision>=0.14.1
scikit-learn>=1.2.2
